+++ src/DataGrid.tsx
@@ -142,7 +142,7 @@ export interface DataGridProps<R, K extends keyof R> {
  * <DataGrid columns={columns} rows={rows} />
 */
 function DataGrid<R, K extends keyof R>({
-  rowKey = 'id' as K,
+  rowKey,
   rowHeight = 35,
   headerRowHeight = rowHeight,
   headerFiltersHeight = 45,
@@ -276,7 +276,7 @@ function DataGrid<R, K extends keyof R>({
               scrollLeft={nonStickyScrollLeft}
             />
             {enableFilters && (
-              <FilterRow<R, K>
+              <FilterRow<R>
                 height={headerFiltersHeight}
                 width={columnMetrics.totalColumnWidth + getScrollbarSize()}
                 lastFrozenColumnIndex={columnMetrics.lastFrozenColumnIndex}
@@ -324,5 +324,5 @@ function DataGrid<R, K extends keyof R>({
 }
 
 export default forwardRef(
-  DataGrid as React.RefForwardingComponent<DataGridHandle, DataGridProps<{ [key: string]: unknown }, string>>
+  DataGrid as React.RefForwardingComponent<DataGridHandle>
 ) as <R, K extends keyof R>(props: DataGridProps<R, K> & { ref?: React.Ref<DataGridHandle> }) => JSX.Element;